[AUII-05] Se implementa logica de vci para detectar operadores, identificadores y constantes. Se reestructura el proyecto
