--- HelloWorld.java
@@ -1,11 +1,0 @@
-import javax.swing.JOptionPane;
-
-public class HelloWorld
-{
-    public static void main(String[] args)
-    {
-        String message = "Hello World!"; 
-
-        JOptionPane.showMessageDialog(null, message);
-    }
-}
