--- codigoIntermedio/App.java
@@ -1,9 +1,0 @@
-package codigoIntermedio;
-
-public class App 
-{
-	public static void main(String[] args) 
-	{
-		
-	}
-}
