+++ files/vci.txt
@@ -1,0 +1,9 @@
+resultado%	-52
+centi%	-52
+9	-61
+*	-21
+5	-61
+/	-22
+32	-61
++	-24
+=	-26
