+++ models/Variable.java
@@ -1,4 +1,4 @@
-package semantica;
+package models;
 
 import java.util.ArrayList;
 import java.util.List;
