+++ services/CodigoIntermedioService.java
@@ -1,0 +1,4 @@
+package services;
+
+public class CodigoIntermedioService extends BaseService
+{}
