+++ tokensPractica5.txt
@@ -10,26 +10,16 @@ real	-12	-1	5
 auxiliar%	-52	-2	5
 ;	-75	-1	5
 inicio	-2	-1	6
-leer	-4	-1	7
+resultado%	-52	-2	7
+=	-26	-1	7
 (	-73	-1	7
 centi%	-52	-2	7
+*	-21	-1	7
+9	-61	-1	7
+/	-22	-1	7
+5	-61	-1	7
 )	-74	-1	7
++	-24	-1	7
+32	-61	-1	7
 ;	-75	-1	7
-resultado%	-52	-2	8
-=	-26	-1	8
-(	-73	-1	8
-centi%	-52	-2	8
-*	-21	-1	8
-9	-61	-1	8
-/	-22	-1	8
-5	-61	-1	8
-)	-74	-1	8
-+	-24	-1	8
-32	-61	-1	8
-;	-75	-1	8
-escribir	-5	-1	9
-(	-73	-1	9
-resultado%	-52	-2	9
-)	-74	-1	9
-;	-75	-1	9
-fin	-3	-1	10
+fin	-3	-1	8
